requirements.txt
streamlit 
requests 
beautifulsoup4 
psycopg2-binary 
pandas 
numpy
python-dotenv